+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Remove all extension defined key bindings
 * Remove editor/context menu commands to navigate to the next/previous error/warning
 * Rename editor/context menu item "Load Project" to "Load Directory"
+* Fixed scanning for dead code, generating documentation, and generating diagrams to load the current directory loader file first
 
 ## [0.8.4]